packages/sdk-python/pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aex-sdk"
version = "1.2.0a3"
description = "Python SDK for AEX — the Agent Exchange Protocol."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "Apache-2.0" }
authors = [{ name = "Icaro Holding", email = "oss@spize.io" }]
keywords = ["aex", "agent", "protocol", "identity", "transfer", "p2p"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "httpx>=0.27",
    "cryptography>=42",
    "dnspython>=2.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.23",
    "ruff>=0.3",
    "mypy>=1.8",
]

[project.urls]
Homepage = "https://spize.io"
Repository = "https://github.com/icaroholding/aex"
Documentation = "https://github.com/icaroholding/aex/tree/master/docs"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
aex_sdk = ["py.typed"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
